More efficient texture atlas loading
+ Torch animated sprite
+ Removed unused variables
+ BlockData & BlockEntityData now loads the texture atlases without
creating additional GPU side textures (by loading them as Pixmaps
instead of Texture2DPtrs)
+ Removed some warnings

+++ Source/Gui/LineEdit.cpp
@@ -156,8 +156,8 @@ void LineEdit::keyEvent(const KeyEvent & event)
 			if(event.isModifier(KeyEvent::CONTROL))
 			{
 				// Move cursor to the left of the next word
-				while(m_cursorPos <= m_text.size() && m_text[m_cursorPos++] != ' ');
-				while(m_cursorPos <= m_text.size() && m_text[m_cursorPos++] == ' ');
+				while((uint) m_cursorPos <= m_text.size() && m_text[m_cursorPos++] != ' ');
+				while((uint) m_cursorPos <= m_text.size() && m_text[m_cursorPos++] == ' ');
 				m_cursorPos--;
 			}
 			else
